rename folder to config
